requirements.txt
openai=1.61.0
httpx=0.28.1
python-dotenv=1.0.1
langgraph=0.2.69
langchain_community=0.3.16
langchain_openai=0.3.3
matplotlib=3.10.0
langgraph-checkpoint-sqlite=2.0.3